@@ -1,21 +1,25 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp5.0</TargetFramework>
     <LangVersion>latest</LangVersion>
+    <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ACMESharpCore" Version="2.0.0.49-beta1" />
-    <PackageReference Include="Docker.DotNet" Version="3.125.2" />
-    <PackageReference Include="KdSoft.FlatBuffers" Version="1.9.3" />
-    <PackageReference Include="Nito.AsyncEx" Version="5.0.0-pre-05" />
-    <PackageReference Include="RBush" Version="2.0.31" />
+    <PackageReference Include="FlatBuffers.Core" Version="1.12.0" />
+    <!--  <PackageReference Include="Docker.DotNet" Version="3.125.4" /> -->
+    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.7" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.2" />
+    <PackageReference Include="Nito.AsyncEx" Version="5.1.0" />
+    <PackageReference Include="RBush" Version="2.0.46" />
     <PackageReference Include="System.Numerics.Vectors" Version="4.5.0" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Discord.Net" Version="2.0.0-beta2-01039" />
-    <PackageReference Include="DnsClient" Version="1.1.1" />
-    <PackageReference Include="Makaretu.Dns.Unicast" Version="0.5.0" />
+    <PackageReference Include="Discord.Net" Version="2.2.0" />
+    <PackageReference Include="NEST" Version="7.10.1" />
+    <PackageReference Include="NEST.JsonNetSerializer" Version="7.10.1" />
+    <!-- <PackageReference Include="DnsClient" Version="1.3.2" /> -->
+    <!-- <PackageReference Include="Makaretu.Dns.Unicast" Version="0.11.1" /> -->
   </ItemGroup>
 
   <ItemGroup>
@@ -23,7 +27,6 @@
     <ProjectReference Include="..\Game.API.Common\Game.API.Common.csproj" />
     <ProjectReference Include="..\Game.Engine.Networking.FlatBuffers\Game.Engine.Networking.FlatBuffers.csproj" />
     <ProjectReference Include="..\Game.Util\Game.Util.csproj" />
-    <ProjectReference Include="..\PKISharp.SimplePKI\PKISharp.SimplePKI.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -36,9 +39,5 @@
     <None Update="Dockerfile">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="wwwroot\img\worlds\" />
   </ItemGroup>
 </Project>
