@@ -1,12 +1,8 @@
 ﻿namespace Game.Engine
 {
     using Game.Engine.ChatBot;
-    using Game.Engine.Crypto.LetsEncrypt;
-    using Game.Engine.Hosting;
     using Microsoft.AspNetCore;
     using Microsoft.AspNetCore.Hosting;
-    using Microsoft.Extensions.Configuration;
-    using System;
     using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
@@ -23,7 +19,7 @@
 
         public async static Task Main(string[] args)
         {
-            var host = (await CreateWebHostBuilderAsync(args)).Build();
+            var host = CreateWebHostBuilder(args).Build();
 
             if (cts.IsCancellationRequested)
                 return;
@@ -34,50 +30,23 @@
             await host.RunAsync(cts.Token);
         }
 
-        public async static Task<IWebHostBuilder> CreateWebHostBuilderAsync(string[] args)
+        public static IWebHostBuilder CreateWebHostBuilder(string[] args)
         {
             ThreadPool.SetMinThreads(50, 50);
 
             var builder = WebHost.CreateDefaultBuilder(args);
 
-            var config = new GameConfiguration();
-            Configuration<GameConfiguration>.Load("config", instance: config);
+            var configContext = Configuration<GameConfiguration>.Load("config");
 
-            if (Environment.GetEnvironmentVariable("GAME_DOCKER_UPGRADE") != null)
-            {
-                var oldID = Environment.GetEnvironmentVariable("GAME_DOCKER_UPGRADE_OLD");
-                var newID = Environment.GetEnvironmentVariable("GAME_DOCKER_UPGRADE_NEW");
-
-                await DockerUpgrade.FinalSwitchAsync(config, oldID, newID);
-                Program.Abort();
-            }
+            builder.UseConfiguration(configContext.ConfigurationRoot);
 
             var port = System.Environment.GetEnvironmentVariable("PORT");
-
             if (!string.IsNullOrEmpty(port))
-            {
                 builder.UseUrls($"http://*:{port}");
-            }
-            else
-                builder.UseConfiguration(new ConfigurationBuilder()
-                    .AddJsonFile("hosting.json", optional: true)
-                    .Build()
-                );
 
             builder
                 .UseWebRoot("wwwroot/dist")
                 .UseStartup<Startup>();
-
-
-            if (config.LetsEncryptEnabled && config.RegistryEnabled)
-                // Full Form with access to All Options:
-                builder.AddAcmeServices(new AcmeOptions
-                {
-                    AcmeRootDir = "_IGNORE/_acmesharp",
-                    AccountContactEmails = new[] { "info@daud.io" },
-                    AcceptTermsOfService = true,
-                    CertificateKeyAlgor = "rsa",
-                });
 
             return builder;
         }
